backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
easyocr==1.7.0
pyspellchecker==0.7.2
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
psutil==5.9.6
pytest==7.4.3
groq==0.11.0
